Added ability to have more than one epoch in a training pass
Changed train_mapper.py to explicitly use an environment variable, T,
to set the number of iterations over the training set. This has the
consequence of requiring the storage of all training instances in
memory to support iteration; this will have to be taken into account in
the provisioning of resources for execution. The main loop is closer in
implementation to the pseudocode on page 6 in William Cohen's notes on
stochastic gradient descent [1], and the learning rate is now decreased
according to the formula in the second sentence of page 7 of that
reference, as opposed to the previously utilized alternative of
1/(number of examples).